plot_reproductive_epi Function

This function generates a plot of the reproductive number over time

plot_reproductive_epi(model)

Arguments

model

The model object

Value

A plot displaying the reproductive number for the model over the course of the simulation

Examples

library(epiworldR) # for ModelSEIRCONN function
model <- ModelSEIRCONN("COVID-19", n = 1000, prevalence = 0.05,
                       contact_rate = 4, transmission_rate = 0.1,
                       incubation_days = 7, recovery_rate = 0.14)
run(model, ndays = 100, seed = 123)
#> _________________________________________________________________________
#> Running the model...
#> ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| done.
#>  done.
plot_reproductive_epi(model)