plot_epi Function

This function generates a plot of the model states over time

plot_epi(model, mark_max)

Arguments

model

The number of individuals in the population.

mark_max

The state which will have a mark at the peak

Value

A plot displaying each state from the model over the course of the simulation

Examples

library(epiworldR) # for ModelSEIRCONN function
model <- ModelSEIRCONN("COVID-19", n = 1000, prevalence = 0.05,
                       contact_rate = 4, transmission_rate = 0.1,
                       incubation_days = 7, recovery_rate = 0.14)
run(model, ndays = 100, seed = 123)
#> _________________________________________________________________________
#> Running the model...
#> ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| done.
#>  done.
plot_epi(model, mark_max = "Infected")