Population generator

This function generates a population matrix with specified characteristics.

pop_generator(
  n,
  prop_hispanic = 0.5,
  prop_female = 0.5,
  prop_19_59_60plus = c(0.3, 0.6)
)

Arguments

n

The number of individuals in the population.

prop_hispanic

The proportion of individuals who are Hispanic. Default is 0.5.

prop_female

The proportion of individuals who are female. Default is 0.5.

prop_19_59_60plus

A vector of length 3 representing the proportions of individuals in the age groups 0-19, 20-59, and 60+. Default is c(0.5, 0.3, 0.2).

Value

A matrix representing the generated population, with columns for age groups (0-19, 20-59, 60+), NotHispanic, and Female.

See also

Examples

pop_generator(n = 1000, prop_hispanic = .5, prop_female = .5,
              prop_19_59_60plus = c(.3, .6))
#>         0-19 20-59 60+ NotHispanic Female
#>    [1,]    1     0   0           0      1
#>    [2,]    0     1   0           1      1
#>    [3,]    1     0   0           0      0
#>    [4,]    1     0   0           0      1
#>    [5,]    0     0   1           1      0
#>    [6,]    0     1   0           0      1
#>    [7,]    1     0   0           1      1
#>    [8,]    0     1   0           1      1
#>    [9,]    0     1   0           1      1
#>   [10,]    1     0   0           1      0
#>   [11,]    0     1   0           1      1
#>   [12,]    0     1   0           1      0
#>   [13,]    0     1   0           0      0
#>   [14,]    0     0   1           1      0
#>   [15,]    1     0   0           0      1
#>   [16,]    0     0   1           0      0
#>   [17,]    0     0   1           0      1
#>   [18,]    0     0   1           1      1
#>   [19,]    1     0   0           1      1
#>   [20,]    1     0   0           1      0
#>   [21,]    0     1   0           0      1
#>   [22,]    0     1   0           0      0
#>   [23,]    1     0   0           1      1
#>   [24,]    0     1   0           0      1
#>   [25,]    1     0   0           1      1
#>   [26,]    0     1   0           0      0
#>   [27,]    0     1   0           1      0
#>   [28,]    0     0   1           1      1
#>   [29,]    0     0   1           0      0
#>   [30,]    1     0   0           1      0
#>   [31,]    1     0   0           0      0
#>   [32,]    0     1   0           1      0
#>   [33,]    1     0   0           1      1
#>   [34,]    0     0   1           0      1
#>   [35,]    0     1   0           0      0
#>   [36,]    1     0   0           1      0
#>   [37,]    0     0   1           0      0
#>   [38,]    0     0   1           1      0
#>   [39,]    0     0   1           1      0
#>   [40,]    0     0   1           0      1
#>   [41,]    0     1   0           0      1
#>   [42,]    0     1   0           1      1
#>   [43,]    0     0   1           0      0
#>   [44,]    0     0   1           1      0
#>   [45,]    0     0   1           1      0
#>   [46,]    0     0   1           1      0
#>   [47,]    0     1   0           1      1
#>   [48,]    0     0   1           0      1
#>   [49,]    1     0   0           0      1
#>   [50,]    0     0   1           0      0
#>   [51,]    0     1   0           0      1
#>   [52,]    1     0   0           0      0
#>   [53,]    0     0   1           0      1
#>   [54,]    0     1   0           0      0
#>   [55,]    0     0   1           1      1
#>   [56,]    0     0   1           0      0
#>   [57,]    1     0   0           1      0
#>   [58,]    1     0   0           1      0
#>   [59,]    0     0   1           0      0
#>   [60,]    0     1   0           0      1
#>   [61,]    0     0   1           1      0
#>   [62,]    0     0   1           1      0
#>   [63,]    0     0   1           1      0
#>   [64,]    1     0   0           1      1
#>   [65,]    0     1   0           0      1
#>   [66,]    1     0   0           0      1
#>   [67,]    1     0   0           1      1
#>   [68,]    1     0   0           0      0
#>   [69,]    0     1   0           0      0
#>   [70,]    1     0   0           0      1
#>   [71,]    0     1   0           0      0
#>   [72,]    1     0   0           0      0
#>   [73,]    0     0   1           0      1
#>   [74,]    0     1   0           0      0
#>   [75,]    0     0   1           0      0
#>   [76,]    0     0   1           1      0
#>   [77,]    0     1   0           1      1
#>   [78,]    0     0   1           1      0
#>   [79,]    0     0   1           1      1
#>   [80,]    0     0   1           1      1
#>   [81,]    0     1   0           1      0
#>   [82,]    0     1   0           1      0
#>   [83,]    1     0   0           1      1
#>   [84,]    0     0   1           0      1
#>   [85,]    0     1   0           1      0
#>   [86,]    1     0   0           1      1
#>   [87,]    1     0   0           1      0
#>   [88,]    1     0   0           0      0
#>   [89,]    0     0   1           1      1
#>   [90,]    0     0   1           0      1
#>   [91,]    0     1   0           1      1
#>   [92,]    0     0   1           1      1
#>   [93,]    0     1   0           1      1
#>   [94,]    0     0   1           0      1
#>   [95,]    0     0   1           0      0
#>   [96,]    1     0   0           0      0
#>   [97,]    0     0   1           0      1
#>   [98,]    0     1   0           1      0
#>   [99,]    0     1   0           1      1
#>  [100,]    0     1   0           0      0
#>  [101,]    0     0   1           0      1
#>  [102,]    0     1   0           1      1
#>  [103,]    1     0   0           0      0
#>  [104,]    0     1   0           0      0
#>  [105,]    1     0   0           0      0
#>  [106,]    1     0   0           0      1
#>  [107,]    0     1   0           0      1
#>  [108,]    0     1   0           0      0
#>  [109,]    0     0   1           0      0
#>  [110,]    1     0   0           1      0
#>  [111,]    0     0   1           0      1
#>  [112,]    0     0   1           1      1
#>  [113,]    1     0   0           1      1
#>  [114,]    1     0   0           1      0
#>  [115,]    0     0   1           1      1
#>  [116,]    0     1   0           1      0
#>  [117,]    1     0   0           0      1
#>  [118,]    0     1   0           1      0
#>  [119,]    0     1   0           0      0
#>  [120,]    0     1   0           1      0
#>  [121,]    0     0   1           1      0
#>  [122,]    0     0   1           0      1
#>  [123,]    0     0   1           1      1
#>  [124,]    0     0   1           1      0
#>  [125,]    1     0   0           1      1
#>  [126,]    0     0   1           0      0
#>  [127,]    0     0   1           0      0
#>  [128,]    0     0   1           1      1
#>  [129,]    0     1   0           0      0
#>  [130,]    0     1   0           0      1
#>  [131,]    1     0   0           0      1
#>  [132,]    0     1   0           0      1
#>  [133,]    1     0   0           0      1
#>  [134,]    0     1   0           0      0
#>  [135,]    0     1   0           0      0
#>  [136,]    1     0   0           1      1
#>  [137,]    1     0   0           1      0
#>  [138,]    1     0   0           0      0
#>  [139,]    0     1   0           1      0
#>  [140,]    0     0   1           0      0
#>  [141,]    0     1   0           1      1
#>  [142,]    0     0   1           1      1
#>  [143,]    0     0   1           0      1
#>  [144,]    1     0   0           1      1
#>  [145,]    0     0   1           1      1
#>  [146,]    0     0   1           0      1
#>  [147,]    0     0   1           0      1
#>  [148,]    0     0   1           1      0
#>  [149,]    1     0   0           1      0
#>  [150,]    1     0   0           0      1
#>  [151,]    0     1   0           0      0
#>  [152,]    0     0   1           0      0
#>  [153,]    0     0   1           0      1
#>  [154,]    0     0   1           0      1
#>  [155,]    0     0   1           1      1
#>  [156,]    0     1   0           1      1
#>  [157,]    0     0   1           0      0
#>  [158,]    0     1   0           0      1
#>  [159,]    0     0   1           1      1
#>  [160,]    0     1   0           1      0
#>  [161,]    0     0   1           1      1
#>  [162,]    0     1   0           1      0
#>  [163,]    0     0   1           1      0
#>  [164,]    0     0   1           1      0
#>  [165,]    0     1   0           1      0
#>  [166,]    1     0   0           0      1
#>  [167,]    0     0   1           1      0
#>  [168,]    0     1   0           1      0
#>  [169,]    0     0   1           0      1
#>  [170,]    0     1   0           1      1
#>  [171,]    0     0   1           1      1
#>  [172,]    1     0   0           0      1
#>  [173,]    1     0   0           0      1
#>  [174,]    0     1   0           1      0
#>  [175,]    0     1   0           0      1
#>  [176,]    0     0   1           0      1
#>  [177,]    0     1   0           1      0
#>  [178,]    1     0   0           1      1
#>  [179,]    0     1   0           0      0
#>  [180,]    1     0   0           1      0
#>  [181,]    0     0   1           1      0
#>  [182,]    1     0   0           0      1
#>  [183,]    0     0   1           0      1
#>  [184,]    0     1   0           0      0
#>  [185,]    0     1   0           1      1
#>  [186,]    0     0   1           1      0
#>  [187,]    0     1   0           0      0
#>  [188,]    1     0   0           0      1
#>  [189,]    1     0   0           0      0
#>  [190,]    0     0   1           1      0
#>  [191,]    0     0   1           0      1
#>  [192,]    1     0   0           1      0
#>  [193,]    0     1   0           0      0
#>  [194,]    1     0   0           1      0
#>  [195,]    0     1   0           1      0
#>  [196,]    0     1   0           0      0
#>  [197,]    0     1   0           1      0
#>  [198,]    0     0   1           0      1
#>  [199,]    0     1   0           1      1
#>  [200,]    0     0   1           0      0
#>  [201,]    1     0   0           0      0
#>  [202,]    0     1   0           1      0
#>  [203,]    0     1   0           0      1
#>  [204,]    0     1   0           1      1
#>  [205,]    1     0   0           0      0
#>  [206,]    0     0   1           1      1
#>  [207,]    0     0   1           1      0
#>  [208,]    0     0   1           0      0
#>  [209,]    0     0   1           0      1
#>  [210,]    0     1   0           1      1
#>  [211,]    0     0   1           0      0
#>  [212,]    0     0   1           1      1
#>  [213,]    0     1   0           0      1
#>  [214,]    0     0   1           1      0
#>  [215,]    1     0   0           0      1
#>  [216,]    0     0   1           1      1
#>  [217,]    0     1   0           0      0
#>  [218,]    1     0   0           1      0
#>  [219,]    1     0   0           0      1
#>  [220,]    0     0   1           1      1
#>  [221,]    1     0   0           0      0
#>  [222,]    1     0   0           1      0
#>  [223,]    0     1   0           0      1
#>  [224,]    0     0   1           0      0
#>  [225,]    0     0   1           1      1
#>  [226,]    0     1   0           1      1
#>  [227,]    0     1   0           1      1
#>  [228,]    0     1   0           0      0
#>  [229,]    1     0   0           0      1
#>  [230,]    0     1   0           1      0
#>  [231,]    0     1   0           1      0
#>  [232,]    0     1   0           1      1
#>  [233,]    0     0   1           0      0
#>  [234,]    0     0   1           1      1
#>  [235,]    0     0   1           1      1
#>  [236,]    0     0   1           1      0
#>  [237,]    1     0   0           0      1
#>  [238,]    0     0   1           0      1
#>  [239,]    1     0   0           1      0
#>  [240,]    0     1   0           1      1
#>  [241,]    0     1   0           1      0
#>  [242,]    0     0   1           0      1
#>  [243,]    0     1   0           1      0
#>  [244,]    0     0   1           0      0
#>  [245,]    1     0   0           1      0
#>  [246,]    0     0   1           0      1
#>  [247,]    1     0   0           1      0
#>  [248,]    1     0   0           1      1
#>  [249,]    1     0   0           0      1
#>  [250,]    0     0   1           0      0
#>  [251,]    0     0   1           0      0
#>  [252,]    0     1   0           0      1
#>  [253,]    0     0   1           1      0
#>  [254,]    0     1   0           1      1
#>  [255,]    1     0   0           0      1
#>  [256,]    0     0   1           1      0
#>  [257,]    0     1   0           1      1
#>  [258,]    0     1   0           0      1
#>  [259,]    1     0   0           0      0
#>  [260,]    1     0   0           1      0
#>  [261,]    1     0   0           0      0
#>  [262,]    0     1   0           1      1
#>  [263,]    1     0   0           1      1
#>  [264,]    0     1   0           0      0
#>  [265,]    0     1   0           1      1
#>  [266,]    0     0   1           0      1
#>  [267,]    0     0   1           0      0
#>  [268,]    0     0   1           1      1
#>  [269,]    0     1   0           0      1
#>  [270,]    1     0   0           1      1
#>  [271,]    0     0   1           0      1
#>  [272,]    0     0   1           1      1
#>  [273,]    0     0   1           0      1
#>  [274,]    1     0   0           0      0
#>  [275,]    1     0   0           0      0
#>  [276,]    1     0   0           0      0
#>  [277,]    0     1   0           0      1
#>  [278,]    0     0   1           1      1
#>  [279,]    0     1   0           1      1
#>  [280,]    1     0   0           1      1
#>  [281,]    0     0   1           0      0
#>  [282,]    0     0   1           0      1
#>  [283,]    0     0   1           0      0
#>  [284,]    0     0   1           0      0
#>  [285,]    0     0   1           1      1
#>  [286,]    0     0   1           0      0
#>  [287,]    0     0   1           0      1
#>  [288,]    0     0   1           1      0
#>  [289,]    0     1   0           0      0
#>  [290,]    0     0   1           0      1
#>  [291,]    0     0   1           0      0
#>  [292,]    0     0   1           1      1
#>  [293,]    1     0   0           0      1
#>  [294,]    1     0   0           1      0
#>  [295,]    1     0   0           1      1
#>  [296,]    1     0   0           0      1
#>  [297,]    0     0   1           0      0
#>  [298,]    0     0   1           0      0
#>  [299,]    1     0   0           0      0
#>  [300,]    1     0   0           1      1
#>  [301,]    0     0   1           1      0
#>  [302,]    1     0   0           0      1
#>  [303,]    1     0   0           0      0
#>  [304,]    0     1   0           0      0
#>  [305,]    0     1   0           0      1
#>  [306,]    0     0   1           1      0
#>  [307,]    0     0   1           1      0
#>  [308,]    0     1   0           0      0
#>  [309,]    0     1   0           1      1
#>  [310,]    0     0   1           1      0
#>  [311,]    0     1   0           1      1
#>  [312,]    1     0   0           0      1
#>  [313,]    0     0   1           1      0
#>  [314,]    0     0   1           0      0
#>  [315,]    1     0   0           1      0
#>  [316,]    1     0   0           0      1
#>  [317,]    0     0   1           1      0
#>  [318,]    0     0   1           0      1
#>  [319,]    1     0   0           1      0
#>  [320,]    1     0   0           0      0
#>  [321,]    0     0   1           1      0
#>  [322,]    0     0   1           0      1
#>  [323,]    1     0   0           1      0
#>  [324,]    0     0   1           0      1
#>  [325,]    0     0   1           1      1
#>  [326,]    1     0   0           1      1
#>  [327,]    0     0   1           1      1
#>  [328,]    0     0   1           1      1
#>  [329,]    1     0   0           0      0
#>  [330,]    0     1   0           0      1
#>  [331,]    0     0   1           1      1
#>  [332,]    1     0   0           0      0
#>  [333,]    1     0   0           0      0
#>  [334,]    0     0   1           1      0
#>  [335,]    0     1   0           1      1
#>  [336,]    0     1   0           0      1
#>  [337,]    0     1   0           0      0
#>  [338,]    0     1   0           0      0
#>  [339,]    1     0   0           0      0
#>  [340,]    0     1   0           1      0
#>  [341,]    0     0   1           1      1
#>  [342,]    0     1   0           1      0
#>  [343,]    0     0   1           1      0
#>  [344,]    0     1   0           0      1
#>  [345,]    0     0   1           0      1
#>  [346,]    1     0   0           1      0
#>  [347,]    0     0   1           0      1
#>  [348,]    0     0   1           0      1
#>  [349,]    0     1   0           0      1
#>  [350,]    0     0   1           1      0
#>  [351,]    1     0   0           1      0
#>  [352,]    0     1   0           1      1
#>  [353,]    0     0   1           1      0
#>  [354,]    0     1   0           1      1
#>  [355,]    1     0   0           1      1
#>  [356,]    0     1   0           1      1
#>  [357,]    0     1   0           0      1
#>  [358,]    0     0   1           1      1
#>  [359,]    1     0   0           1      1
#>  [360,]    0     0   1           0      0
#>  [361,]    0     0   1           1      0
#>  [362,]    1     0   0           0      1
#>  [363,]    0     0   1           1      0
#>  [364,]    0     0   1           1      1
#>  [365,]    1     0   0           1      1
#>  [366,]    0     0   1           1      1
#>  [367,]    0     0   1           0      0
#>  [368,]    0     1   0           1      0
#>  [369,]    0     1   0           1      0
#>  [370,]    0     1   0           1      0
#>  [371,]    0     0   1           1      1
#>  [372,]    1     0   0           1      1
#>  [373,]    0     1   0           0      1
#>  [374,]    0     1   0           0      1
#>  [375,]    1     0   0           0      1
#>  [376,]    1     0   0           1      0
#>  [377,]    0     0   1           0      1
#>  [378,]    0     0   1           1      1
#>  [379,]    1     0   0           0      0
#>  [380,]    0     1   0           1      1
#>  [381,]    1     0   0           1      1
#>  [382,]    0     0   1           0      0
#>  [383,]    1     0   0           1      1
#>  [384,]    0     1   0           0      0
#>  [385,]    1     0   0           1      1
#>  [386,]    0     0   1           1      0
#>  [387,]    0     0   1           0      0
#>  [388,]    0     0   1           1      1
#>  [389,]    0     0   1           1      0
#>  [390,]    1     0   0           0      0
#>  [391,]    0     0   1           0      1
#>  [392,]    1     0   0           1      1
#>  [393,]    1     0   0           1      1
#>  [394,]    0     0   1           1      0
#>  [395,]    0     0   1           0      0
#>  [396,]    0     0   1           0      1
#>  [397,]    0     0   1           1      0
#>  [398,]    0     0   1           1      0
#>  [399,]    1     0   0           1      0
#>  [400,]    1     0   0           1      0
#>  [401,]    0     0   1           1      0
#>  [402,]    1     0   0           1      1
#>  [403,]    0     1   0           1      0
#>  [404,]    0     0   1           1      0
#>  [405,]    0     1   0           0      1
#>  [406,]    1     0   0           0      1
#>  [407,]    0     0   1           0      0
#>  [408,]    0     0   1           1      0
#>  [409,]    0     1   0           0      1
#>  [410,]    0     0   1           0      0
#>  [411,]    1     0   0           0      0
#>  [412,]    0     0   1           0      1
#>  [413,]    0     1   0           1      1
#>  [414,]    0     0   1           1      0
#>  [415,]    0     0   1           0      1
#>  [416,]    1     0   0           0      1
#>  [417,]    0     0   1           1      1
#>  [418,]    0     1   0           0      0
#>  [419,]    0     0   1           1      1
#>  [420,]    0     1   0           0      0
#>  [421,]    0     1   0           0      0
#>  [422,]    0     1   0           1      1
#>  [423,]    0     0   1           0      1
#>  [424,]    1     0   0           1      0
#>  [425,]    0     1   0           1      1
#>  [426,]    0     0   1           1      1
#>  [427,]    0     0   1           0      0
#>  [428,]    0     0   1           0      1
#>  [429,]    1     0   0           1      0
#>  [430,]    1     0   0           0      0
#>  [431,]    0     1   0           1      1
#>  [432,]    0     0   1           0      1
#>  [433,]    1     0   0           0      0
#>  [434,]    0     0   1           1      1
#>  [435,]    0     1   0           1      0
#>  [436,]    0     0   1           0      1
#>  [437,]    0     1   0           0      0
#>  [438,]    0     0   1           0      1
#>  [439,]    0     0   1           0      0
#>  [440,]    0     0   1           1      0
#>  [441,]    0     1   0           0      1
#>  [442,]    1     0   0           1      1
#>  [443,]    0     0   1           1      1
#>  [444,]    1     0   0           0      1
#>  [445,]    1     0   0           0      0
#>  [446,]    1     0   0           1      1
#>  [447,]    0     0   1           1      1
#>  [448,]    0     0   1           0      1
#>  [449,]    0     1   0           1      0
#>  [450,]    0     1   0           0      0
#>  [451,]    0     0   1           0      1
#>  [452,]    0     0   1           0      1
#>  [453,]    0     0   1           1      0
#>  [454,]    0     0   1           0      0
#>  [455,]    1     0   0           0      0
#>  [456,]    1     0   0           1      0
#>  [457,]    1     0   0           0      1
#>  [458,]    0     0   1           0      0
#>  [459,]    0     0   1           1      1
#>  [460,]    1     0   0           0      1
#>  [461,]    0     1   0           0      0
#>  [462,]    0     1   0           0      0
#>  [463,]    0     0   1           1      1
#>  [464,]    0     1   0           1      1
#>  [465,]    0     0   1           0      1
#>  [466,]    0     0   1           0      1
#>  [467,]    0     1   0           1      1
#>  [468,]    0     0   1           1      1
#>  [469,]    1     0   0           1      1
#>  [470,]    0     0   1           0      1
#>  [471,]    1     0   0           1      1
#>  [472,]    0     1   0           0      1
#>  [473,]    1     0   0           1      0
#>  [474,]    0     1   0           1      1
#>  [475,]    0     1   0           0      1
#>  [476,]    0     1   0           1      0
#>  [477,]    0     0   1           0      1
#>  [478,]    0     0   1           0      0
#>  [479,]    1     0   0           0      1
#>  [480,]    0     0   1           0      1
#>  [481,]    1     0   0           1      1
#>  [482,]    0     0   1           0      0
#>  [483,]    0     1   0           0      1
#>  [484,]    1     0   0           0      0
#>  [485,]    0     0   1           1      1
#>  [486,]    0     0   1           0      1
#>  [487,]    0     1   0           1      0
#>  [488,]    0     1   0           0      1
#>  [489,]    1     0   0           0      0
#>  [490,]    1     0   0           1      0
#>  [491,]    0     0   1           1      1
#>  [492,]    0     1   0           0      0
#>  [493,]    1     0   0           0      0
#>  [494,]    0     0   1           1      0
#>  [495,]    1     0   0           0      0
#>  [496,]    0     0   1           0      0
#>  [497,]    0     0   1           1      1
#>  [498,]    0     1   0           0      0
#>  [499,]    1     0   0           1      1
#>  [500,]    0     0   1           0      0
#>  [501,]    0     0   1           0      0
#>  [502,]    0     0   1           1      0
#>  [503,]    0     0   1           1      1
#>  [504,]    0     0   1           0      0
#>  [505,]    0     0   1           0      1
#>  [506,]    0     1   0           0      0
#>  [507,]    0     1   0           1      0
#>  [508,]    1     0   0           1      0
#>  [509,]    0     0   1           0      0
#>  [510,]    0     1   0           1      1
#>  [511,]    0     0   1           0      1
#>  [512,]    1     0   0           0      1
#>  [513,]    0     0   1           1      1
#>  [514,]    1     0   0           1      0
#>  [515,]    0     0   1           1      0
#>  [516,]    0     1   0           1      1
#>  [517,]    1     0   0           1      0
#>  [518,]    0     0   1           1      0
#>  [519,]    1     0   0           1      1
#>  [520,]    0     0   1           0      0
#>  [521,]    0     0   1           1      0
#>  [522,]    0     0   1           0      0
#>  [523,]    0     0   1           1      1
#>  [524,]    0     1   0           0      1
#>  [525,]    0     1   0           1      0
#>  [526,]    1     0   0           1      1
#>  [527,]    0     0   1           1      1
#>  [528,]    1     0   0           0      0
#>  [529,]    0     0   1           1      0
#>  [530,]    1     0   0           0      0
#>  [531,]    0     1   0           0      0
#>  [532,]    0     1   0           0      0
#>  [533,]    1     0   0           1      0
#>  [534,]    0     0   1           1      0
#>  [535,]    1     0   0           1      1
#>  [536,]    0     0   1           1      1
#>  [537,]    1     0   0           1      0
#>  [538,]    0     1   0           0      0
#>  [539,]    0     0   1           1      0
#>  [540,]    1     0   0           1      0
#>  [541,]    0     0   1           0      1
#>  [542,]    1     0   0           1      0
#>  [543,]    0     0   1           0      0
#>  [544,]    1     0   0           0      0
#>  [545,]    1     0   0           1      0
#>  [546,]    0     0   1           1      1
#>  [547,]    1     0   0           1      0
#>  [548,]    1     0   0           0      1
#>  [549,]    0     0   1           0      1
#>  [550,]    0     0   1           0      0
#>  [551,]    0     0   1           1      1
#>  [552,]    0     1   0           1      0
#>  [553,]    1     0   0           1      1
#>  [554,]    0     1   0           1      0
#>  [555,]    0     1   0           1      0
#>  [556,]    0     1   0           0      1
#>  [557,]    0     1   0           1      1
#>  [558,]    0     1   0           1      1
#>  [559,]    0     0   1           0      1
#>  [560,]    0     0   1           0      1
#>  [561,]    1     0   0           1      0
#>  [562,]    1     0   0           0      1
#>  [563,]    0     0   1           1      1
#>  [564,]    0     1   0           0      1
#>  [565,]    0     1   0           1      0
#>  [566,]    0     0   1           1      1
#>  [567,]    1     0   0           0      1
#>  [568,]    0     1   0           1      1
#>  [569,]    1     0   0           0      0
#>  [570,]    0     1   0           1      1
#>  [571,]    1     0   0           1      0
#>  [572,]    0     1   0           0      1
#>  [573,]    0     1   0           1      1
#>  [574,]    1     0   0           1      1
#>  [575,]    1     0   0           0      1
#>  [576,]    0     1   0           1      1
#>  [577,]    0     0   1           1      1
#>  [578,]    0     1   0           1      1
#>  [579,]    0     1   0           1      0
#>  [580,]    1     0   0           0      0
#>  [581,]    1     0   0           1      1
#>  [582,]    1     0   0           1      0
#>  [583,]    0     0   1           1      1
#>  [584,]    1     0   0           0      1
#>  [585,]    0     0   1           0      0
#>  [586,]    0     0   1           0      0
#>  [587,]    0     1   0           1      0
#>  [588,]    1     0   0           1      0
#>  [589,]    0     1   0           1      1
#>  [590,]    0     1   0           0      0
#>  [591,]    0     0   1           0      0
#>  [592,]    1     0   0           0      0
#>  [593,]    0     0   1           0      1
#>  [594,]    0     0   1           1      1
#>  [595,]    1     0   0           1      0
#>  [596,]    0     1   0           0      1
#>  [597,]    0     0   1           1      0
#>  [598,]    1     0   0           0      0
#>  [599,]    0     0   1           1      1
#>  [600,]    0     0   1           0      1
#>  [601,]    0     1   0           1      0
#>  [602,]    0     0   1           1      1
#>  [603,]    0     0   1           0      0
#>  [604,]    0     0   1           0      1
#>  [605,]    1     0   0           1      1
#>  [606,]    0     0   1           0      0
#>  [607,]    1     0   0           1      1
#>  [608,]    0     0   1           0      0
#>  [609,]    0     0   1           1      0
#>  [610,]    0     1   0           0      0
#>  [611,]    0     0   1           0      1
#>  [612,]    0     0   1           1      1
#>  [613,]    1     0   0           0      1
#>  [614,]    0     1   0           0      0
#>  [615,]    1     0   0           1      0
#>  [616,]    1     0   0           0      1
#>  [617,]    1     0   0           1      0
#>  [618,]    1     0   0           1      1
#>  [619,]    0     0   1           1      0
#>  [620,]    1     0   0           1      0
#>  [621,]    1     0   0           1      1
#>  [622,]    1     0   0           1      1
#>  [623,]    0     0   1           0      0
#>  [624,]    0     0   1           1      0
#>  [625,]    0     0   1           0      0
#>  [626,]    0     0   1           0      1
#>  [627,]    1     0   0           1      1
#>  [628,]    0     0   1           0      1
#>  [629,]    0     0   1           0      1
#>  [630,]    1     0   0           0      0
#>  [631,]    1     0   0           0      0
#>  [632,]    0     1   0           1      1
#>  [633,]    1     0   0           0      1
#>  [634,]    0     0   1           0      1
#>  [635,]    0     0   1           0      1
#>  [636,]    1     0   0           1      0
#>  [637,]    1     0   0           1      0
#>  [638,]    1     0   0           1      0
#>  [639,]    0     1   0           0      1
#>  [640,]    0     1   0           1      1
#>  [641,]    1     0   0           0      0
#>  [642,]    1     0   0           0      1
#>  [643,]    0     0   1           0      1
#>  [644,]    0     0   1           1      0
#>  [645,]    0     1   0           0      0
#>  [646,]    1     0   0           0      0
#>  [647,]    0     0   1           0      1
#>  [648,]    0     0   1           1      0
#>  [649,]    0     1   0           1      1
#>  [650,]    1     0   0           1      0
#>  [651,]    1     0   0           0      0
#>  [652,]    0     1   0           1      0
#>  [653,]    0     1   0           1      1
#>  [654,]    1     0   0           1      1
#>  [655,]    1     0   0           0      0
#>  [656,]    1     0   0           1      1
#>  [657,]    0     0   1           0      0
#>  [658,]    0     1   0           1      1
#>  [659,]    0     0   1           0      0
#>  [660,]    1     0   0           0      1
#>  [661,]    1     0   0           1      1
#>  [662,]    0     0   1           1      0
#>  [663,]    1     0   0           0      1
#>  [664,]    0     0   1           0      1
#>  [665,]    0     0   1           1      0
#>  [666,]    1     0   0           0      1
#>  [667,]    0     0   1           1      0
#>  [668,]    0     1   0           1      1
#>  [669,]    0     0   1           1      1
#>  [670,]    0     0   1           0      1
#>  [671,]    0     1   0           0      1
#>  [672,]    0     0   1           1      0
#>  [673,]    0     0   1           0      1
#>  [674,]    0     1   0           1      1
#>  [675,]    0     1   0           0      1
#>  [676,]    1     0   0           0      1
#>  [677,]    0     0   1           0      0
#>  [678,]    0     1   0           0      1
#>  [679,]    0     1   0           1      1
#>  [680,]    1     0   0           1      1
#>  [681,]    1     0   0           0      1
#>  [682,]    1     0   0           0      0
#>  [683,]    1     0   0           1      0
#>  [684,]    1     0   0           1      1
#>  [685,]    0     1   0           1      0
#>  [686,]    0     1   0           0      1
#>  [687,]    0     0   1           0      0
#>  [688,]    0     1   0           0      1
#>  [689,]    0     0   1           0      0
#>  [690,]    0     0   1           1      1
#>  [691,]    0     1   0           1      1
#>  [692,]    0     1   0           1      1
#>  [693,]    0     0   1           0      0
#>  [694,]    0     0   1           0      0
#>  [695,]    0     1   0           0      0
#>  [696,]    0     0   1           0      1
#>  [697,]    0     0   1           0      0
#>  [698,]    0     1   0           1      0
#>  [699,]    0     0   1           1      0
#>  [700,]    1     0   0           0      1
#>  [701,]    0     0   1           1      0
#>  [702,]    0     0   1           1      1
#>  [703,]    1     0   0           0      0
#>  [704,]    0     0   1           0      1
#>  [705,]    0     1   0           0      0
#>  [706,]    0     1   0           1      0
#>  [707,]    0     0   1           0      0
#>  [708,]    0     0   1           0      0
#>  [709,]    0     1   0           0      1
#>  [710,]    0     1   0           1      0
#>  [711,]    0     1   0           1      1
#>  [712,]    0     1   0           1      1
#>  [713,]    1     0   0           0      1
#>  [714,]    0     1   0           1      1
#>  [715,]    0     0   1           0      0
#>  [716,]    0     0   1           0      1
#>  [717,]    0     0   1           0      1
#>  [718,]    1     0   0           0      0
#>  [719,]    0     0   1           1      0
#>  [720,]    0     1   0           0      0
#>  [721,]    1     0   0           1      1
#>  [722,]    1     0   0           0      0
#>  [723,]    1     0   0           1      0
#>  [724,]    0     1   0           1      0
#>  [725,]    1     0   0           0      0
#>  [726,]    1     0   0           1      0
#>  [727,]    0     1   0           0      1
#>  [728,]    1     0   0           1      1
#>  [729,]    0     0   1           1      1
#>  [730,]    0     1   0           0      1
#>  [731,]    1     0   0           1      0
#>  [732,]    0     1   0           1      0
#>  [733,]    1     0   0           0      1
#>  [734,]    0     0   1           0      0
#>  [735,]    0     0   1           0      1
#>  [736,]    1     0   0           1      0
#>  [737,]    1     0   0           0      0
#>  [738,]    0     0   1           0      0
#>  [739,]    0     0   1           0      0
#>  [740,]    1     0   0           1      1
#>  [741,]    0     0   1           0      1
#>  [742,]    0     0   1           0      1
#>  [743,]    0     1   0           1      1
#>  [744,]    0     0   1           1      0
#>  [745,]    0     1   0           1      0
#>  [746,]    1     0   0           0      1
#>  [747,]    0     0   1           0      1
#>  [748,]    0     1   0           0      1
#>  [749,]    1     0   0           1      0
#>  [750,]    0     0   1           0      0
#>  [751,]    0     0   1           0      1
#>  [752,]    0     0   1           0      1
#>  [753,]    0     0   1           0      0
#>  [754,]    0     1   0           0      1
#>  [755,]    0     1   0           1      1
#>  [756,]    0     1   0           0      1
#>  [757,]    1     0   0           1      0
#>  [758,]    0     0   1           1      0
#>  [759,]    0     0   1           0      1
#>  [760,]    0     1   0           0      0
#>  [761,]    0     0   1           1      1
#>  [762,]    0     1   0           1      0
#>  [763,]    0     1   0           1      1
#>  [764,]    0     1   0           1      0
#>  [765,]    0     0   1           0      1
#>  [766,]    0     0   1           0      0
#>  [767,]    0     1   0           1      1
#>  [768,]    0     1   0           0      0
#>  [769,]    0     1   0           0      0
#>  [770,]    1     0   0           0      1
#>  [771,]    1     0   0           0      1
#>  [772,]    0     1   0           1      1
#>  [773,]    1     0   0           0      1
#>  [774,]    0     0   1           1      0
#>  [775,]    0     0   1           1      0
#>  [776,]    0     0   1           1      0
#>  [777,]    0     1   0           1      1
#>  [778,]    0     1   0           0      1
#>  [779,]    1     0   0           1      1
#>  [780,]    0     0   1           1      1
#>  [781,]    1     0   0           1      0
#>  [782,]    0     0   1           0      1
#>  [783,]    0     1   0           0      0
#>  [784,]    1     0   0           1      0
#>  [785,]    0     1   0           1      0
#>  [786,]    0     0   1           0      0
#>  [787,]    0     0   1           1      1
#>  [788,]    0     0   1           0      0
#>  [789,]    1     0   0           0      0
#>  [790,]    1     0   0           1      1
#>  [791,]    0     1   0           1      1
#>  [792,]    0     0   1           0      0
#>  [793,]    0     1   0           0      1
#>  [794,]    1     0   0           1      0
#>  [795,]    0     1   0           1      1
#>  [796,]    0     0   1           0      0
#>  [797,]    0     0   1           0      0
#>  [798,]    0     0   1           0      1
#>  [799,]    0     0   1           0      0
#>  [800,]    0     1   0           1      1
#>  [801,]    0     0   1           0      0
#>  [802,]    0     0   1           1      1
#>  [803,]    0     0   1           0      1
#>  [804,]    0     0   1           0      0
#>  [805,]    1     0   0           1      0
#>  [806,]    0     1   0           0      1
#>  [807,]    0     1   0           0      0
#>  [808,]    1     0   0           0      0
#>  [809,]    0     1   0           0      0
#>  [810,]    0     0   1           1      0
#>  [811,]    0     1   0           0      1
#>  [812,]    0     0   1           0      0
#>  [813,]    1     0   0           0      0
#>  [814,]    1     0   0           1      0
#>  [815,]    0     0   1           1      1
#>  [816,]    0     0   1           0      1
#>  [817,]    0     0   1           1      0
#>  [818,]    0     1   0           0      1
#>  [819,]    0     1   0           1      0
#>  [820,]    0     1   0           0      0
#>  [821,]    0     0   1           1      1
#>  [822,]    0     1   0           1      0
#>  [823,]    1     0   0           0      0
#>  [824,]    1     0   0           0      1
#>  [825,]    0     0   1           0      1
#>  [826,]    0     1   0           1      0
#>  [827,]    0     0   1           0      1
#>  [828,]    0     0   1           0      0
#>  [829,]    1     0   0           0      1
#>  [830,]    0     0   1           0      0
#>  [831,]    1     0   0           0      0
#>  [832,]    0     1   0           0      0
#>  [833,]    1     0   0           1      0
#>  [834,]    0     1   0           1      1
#>  [835,]    0     1   0           1      1
#>  [836,]    1     0   0           0      0
#>  [837,]    0     1   0           1      0
#>  [838,]    0     0   1           0      0
#>  [839,]    0     0   1           0      0
#>  [840,]    1     0   0           1      1
#>  [841,]    1     0   0           1      1
#>  [842,]    0     1   0           1      0
#>  [843,]    0     0   1           0      0
#>  [844,]    0     1   0           1      0
#>  [845,]    0     0   1           1      0
#>  [846,]    1     0   0           0      0
#>  [847,]    0     0   1           1      1
#>  [848,]    0     0   1           0      1
#>  [849,]    0     0   1           1      1
#>  [850,]    0     0   1           0      0
#>  [851,]    1     0   0           0      1
#>  [852,]    1     0   0           1      1
#>  [853,]    1     0   0           1      1
#>  [854,]    0     1   0           0      0
#>  [855,]    0     1   0           0      1
#>  [856,]    1     0   0           0      1
#>  [857,]    0     0   1           1      0
#>  [858,]    0     0   1           0      1
#>  [859,]    1     0   0           0      0
#>  [860,]    0     1   0           0      0
#>  [861,]    0     0   1           0      1
#>  [862,]    0     1   0           1      1
#>  [863,]    0     1   0           0      1
#>  [864,]    0     1   0           0      0
#>  [865,]    0     0   1           1      0
#>  [866,]    0     0   1           1      0
#>  [867,]    0     0   1           0      0
#>  [868,]    0     1   0           1      1
#>  [869,]    0     1   0           0      1
#>  [870,]    0     0   1           1      0
#>  [871,]    0     0   1           0      1
#>  [872,]    1     0   0           1      1
#>  [873,]    0     1   0           1      1
#>  [874,]    1     0   0           1      0
#>  [875,]    0     1   0           0      0
#>  [876,]    0     0   1           1      1
#>  [877,]    0     1   0           1      1
#>  [878,]    1     0   0           0      1
#>  [879,]    0     1   0           1      0
#>  [880,]    0     0   1           0      1
#>  [881,]    1     0   0           1      0
#>  [882,]    0     1   0           1      0
#>  [883,]    0     0   1           1      0
#>  [884,]    1     0   0           1      1
#>  [885,]    1     0   0           0      0
#>  [886,]    0     1   0           1      0
#>  [887,]    0     0   1           1      1
#>  [888,]    0     0   1           1      0
#>  [889,]    0     0   1           0      0
#>  [890,]    0     1   0           0      1
#>  [891,]    1     0   0           0      0
#>  [892,]    0     1   0           1      1
#>  [893,]    0     1   0           0      1
#>  [894,]    1     0   0           0      1
#>  [895,]    1     0   0           0      1
#>  [896,]    0     0   1           1      0
#>  [897,]    0     0   1           0      0
#>  [898,]    0     0   1           1      1
#>  [899,]    1     0   0           0      1
#>  [900,]    1     0   0           1      0
#>  [901,]    0     1   0           1      0
#>  [902,]    1     0   0           1      0
#>  [903,]    0     1   0           0      1
#>  [904,]    0     1   0           0      0
#>  [905,]    0     1   0           0      0
#>  [906,]    1     0   0           0      0
#>  [907,]    1     0   0           0      1
#>  [908,]    1     0   0           0      1
#>  [909,]    0     0   1           0      0
#>  [910,]    0     0   1           1      0
#>  [911,]    0     0   1           0      1
#>  [912,]    1     0   0           0      0
#>  [913,]    0     1   0           0      1
#>  [914,]    0     0   1           0      0
#>  [915,]    1     0   0           1      0
#>  [916,]    0     1   0           1      0
#>  [917,]    0     1   0           0      0
#>  [918,]    0     0   1           0      0
#>  [919,]    0     0   1           1      0
#>  [920,]    1     0   0           0      1
#>  [921,]    0     0   1           1      0
#>  [922,]    0     0   1           0      0
#>  [923,]    1     0   0           1      1
#>  [924,]    0     1   0           0      0
#>  [925,]    0     1   0           1      1
#>  [926,]    0     0   1           0      0
#>  [927,]    0     1   0           1      0
#>  [928,]    0     0   1           0      0
#>  [929,]    1     0   0           0      0
#>  [930,]    1     0   0           0      1
#>  [931,]    0     0   1           1      1
#>  [932,]    1     0   0           1      1
#>  [933,]    0     1   0           1      1
#>  [934,]    1     0   0           1      0
#>  [935,]    0     0   1           1      1
#>  [936,]    0     1   0           1      0
#>  [937,]    1     0   0           0      1
#>  [938,]    1     0   0           1      1
#>  [939,]    0     0   1           0      0
#>  [940,]    1     0   0           0      1
#>  [941,]    0     1   0           1      0
#>  [942,]    0     0   1           1      1
#>  [943,]    0     1   0           0      1
#>  [944,]    0     0   1           1      1
#>  [945,]    0     0   1           0      0
#>  [946,]    0     1   0           1      1
#>  [947,]    0     0   1           0      0
#>  [948,]    0     1   0           0      1
#>  [949,]    0     0   1           1      1
#>  [950,]    0     1   0           1      1
#>  [951,]    0     0   1           0      0
#>  [952,]    1     0   0           1      0
#>  [953,]    0     0   1           1      1
#>  [954,]    0     0   1           0      1
#>  [955,]    1     0   0           1      0
#>  [956,]    0     1   0           1      1
#>  [957,]    0     0   1           0      0
#>  [958,]    0     1   0           0      1
#>  [959,]    0     1   0           1      1
#>  [960,]    0     1   0           0      0
#>  [961,]    0     1   0           1      1
#>  [962,]    0     1   0           1      0
#>  [963,]    0     0   1           1      1
#>  [964,]    1     0   0           1      0
#>  [965,]    0     1   0           1      1
#>  [966,]    0     1   0           1      0
#>  [967,]    0     0   1           1      1
#>  [968,]    1     0   0           0      1
#>  [969,]    0     1   0           0      1
#>  [970,]    1     0   0           1      0
#>  [971,]    0     0   1           0      1
#>  [972,]    1     0   0           1      1
#>  [973,]    0     0   1           0      1
#>  [974,]    0     0   1           0      0
#>  [975,]    0     1   0           0      1
#>  [976,]    1     0   0           1      0
#>  [977,]    0     0   1           1      1
#>  [978,]    0     0   1           0      0
#>  [979,]    0     1   0           1      1
#>  [980,]    0     1   0           0      0
#>  [981,]    1     0   0           1      1
#>  [982,]    1     0   0           1      1
#>  [983,]    0     1   0           1      1
#>  [984,]    1     0   0           0      1
#>  [985,]    1     0   0           1      0
#>  [986,]    0     0   1           1      0
#>  [987,]    0     0   1           1      1
#>  [988,]    0     0   1           0      1
#>  [989,]    0     1   0           1      0
#>  [990,]    0     0   1           1      0
#>  [991,]    0     0   1           1      1
#>  [992,]    0     1   0           0      0
#>  [993,]    0     1   0           1      1
#>  [994,]    0     1   0           1      0
#>  [995,]    1     0   0           0      1
#>  [996,]    0     1   0           1      0
#>  [997,]    0     0   1           0      1
#>  [998,]    1     0   0           1      1
#>  [999,]    0     0   1           1      0
#> [1000,]    0     0   1           1      1