Population generator

This function generates a population matrix with specified characteristics.

pop_generator(
  n,
  prop_hispanic = 0.5,
  prop_female = 0.5,
  prop_19_59_60plus = c(0.3, 0.6)
)

Arguments

n

The number of individuals in the population.

prop_hispanic

The proportion of individuals who are Hispanic. Default is 0.5.

prop_female

The proportion of individuals who are female. Default is 0.5.

prop_19_59_60plus

A vector of length 3 representing the proportions of individuals in the age groups 0-19, 20-59, and 60+. Default is c(0.5, 0.3, 0.2).

Value

A matrix representing the generated population, with columns for age groups (0-19, 20-59, 60+), NotHispanic, and Female.

See also

Examples

pop_generator(n = 1000, prop_hispanic = .5, prop_female = .5,
              prop_19_59_60plus = c(.3, .6))
#>         0-19 20-59 60+ NotHispanic Female
#>    [1,]    1     0   0           0      0
#>    [2,]    1     0   0           0      1
#>    [3,]    0     0   1           1      0
#>    [4,]    0     1   0           0      1
#>    [5,]    1     0   0           1      1
#>    [6,]    0     1   0           1      1
#>    [7,]    0     1   0           1      1
#>    [8,]    1     0   0           1      0
#>    [9,]    0     1   0           1      1
#>   [10,]    0     1   0           1      0
#>   [11,]    0     1   0           0      0
#>   [12,]    0     0   1           1      0
#>   [13,]    1     0   0           0      1
#>   [14,]    0     0   1           0      0
#>   [15,]    0     0   1           0      1
#>   [16,]    0     0   1           1      1
#>   [17,]    1     0   0           1      1
#>   [18,]    1     0   0           1      0
#>   [19,]    0     1   0           0      1
#>   [20,]    0     1   0           0      0
#>   [21,]    1     0   0           1      1
#>   [22,]    0     1   0           0      1
#>   [23,]    1     0   0           1      1
#>   [24,]    0     1   0           0      0
#>   [25,]    0     1   0           1      0
#>   [26,]    0     0   1           1      1
#>   [27,]    0     0   1           0      0
#>   [28,]    1     0   0           1      0
#>   [29,]    1     0   0           0      0
#>   [30,]    0     1   0           1      0
#>   [31,]    1     0   0           1      1
#>   [32,]    0     0   1           0      1
#>   [33,]    0     1   0           0      0
#>   [34,]    1     0   0           1      0
#>   [35,]    0     0   1           0      0
#>   [36,]    0     0   1           1      0
#>   [37,]    0     0   1           1      0
#>   [38,]    0     0   1           0      1
#>   [39,]    0     1   0           0      1
#>   [40,]    0     1   0           1      1
#>   [41,]    0     0   1           0      0
#>   [42,]    0     0   1           1      0
#>   [43,]    0     0   1           1      0
#>   [44,]    0     0   1           1      0
#>   [45,]    0     1   0           1      1
#>   [46,]    0     0   1           0      1
#>   [47,]    1     0   0           0      1
#>   [48,]    0     0   1           0      0
#>   [49,]    0     1   0           0      1
#>   [50,]    1     0   0           0      0
#>   [51,]    0     0   1           0      1
#>   [52,]    0     1   0           0      0
#>   [53,]    0     0   1           1      1
#>   [54,]    0     0   1           0      0
#>   [55,]    1     0   0           1      0
#>   [56,]    1     0   0           1      0
#>   [57,]    0     0   1           0      0
#>   [58,]    0     1   0           0      1
#>   [59,]    0     0   1           1      0
#>   [60,]    0     0   1           1      0
#>   [61,]    0     0   1           1      0
#>   [62,]    1     0   0           1      1
#>   [63,]    0     1   0           0      1
#>   [64,]    1     0   0           0      1
#>   [65,]    1     0   0           1      1
#>   [66,]    1     0   0           0      0
#>   [67,]    0     1   0           0      0
#>   [68,]    1     0   0           0      1
#>   [69,]    0     1   0           0      0
#>   [70,]    1     0   0           0      0
#>   [71,]    0     0   1           0      1
#>   [72,]    0     1   0           0      0
#>   [73,]    0     0   1           0      0
#>   [74,]    0     0   1           1      0
#>   [75,]    0     1   0           1      1
#>   [76,]    0     0   1           1      0
#>   [77,]    0     0   1           1      1
#>   [78,]    0     0   1           1      1
#>   [79,]    0     1   0           1      0
#>   [80,]    0     1   0           1      0
#>   [81,]    1     0   0           1      1
#>   [82,]    0     0   1           0      1
#>   [83,]    0     1   0           1      0
#>   [84,]    1     0   0           1      1
#>   [85,]    1     0   0           1      0
#>   [86,]    1     0   0           0      0
#>   [87,]    0     0   1           1      1
#>   [88,]    0     0   1           0      1
#>   [89,]    0     1   0           1      1
#>   [90,]    0     0   1           1      1
#>   [91,]    0     1   0           1      1
#>   [92,]    0     0   1           0      1
#>   [93,]    0     0   1           0      0
#>   [94,]    1     0   0           0      0
#>   [95,]    0     0   1           0      1
#>   [96,]    0     1   0           1      0
#>   [97,]    0     1   0           1      1
#>   [98,]    0     1   0           0      0
#>   [99,]    0     0   1           0      1
#>  [100,]    0     1   0           1      1
#>  [101,]    1     0   0           0      0
#>  [102,]    0     1   0           0      0
#>  [103,]    1     0   0           0      0
#>  [104,]    1     0   0           0      1
#>  [105,]    0     1   0           0      1
#>  [106,]    0     1   0           0      0
#>  [107,]    0     0   1           0      0
#>  [108,]    1     0   0           1      0
#>  [109,]    0     0   1           0      1
#>  [110,]    0     0   1           1      1
#>  [111,]    1     0   0           1      1
#>  [112,]    1     0   0           1      0
#>  [113,]    0     0   1           1      1
#>  [114,]    0     1   0           1      0
#>  [115,]    1     0   0           0      1
#>  [116,]    0     1   0           1      0
#>  [117,]    0     1   0           0      0
#>  [118,]    0     1   0           1      0
#>  [119,]    0     0   1           1      0
#>  [120,]    0     0   1           0      1
#>  [121,]    0     0   1           1      1
#>  [122,]    0     0   1           1      0
#>  [123,]    1     0   0           1      1
#>  [124,]    0     0   1           0      0
#>  [125,]    0     0   1           0      0
#>  [126,]    0     0   1           1      1
#>  [127,]    0     1   0           0      0
#>  [128,]    0     1   0           0      1
#>  [129,]    1     0   0           0      1
#>  [130,]    0     1   0           0      1
#>  [131,]    1     0   0           0      1
#>  [132,]    0     1   0           0      0
#>  [133,]    0     1   0           0      0
#>  [134,]    1     0   0           1      1
#>  [135,]    1     0   0           1      0
#>  [136,]    1     0   0           0      0
#>  [137,]    0     1   0           1      0
#>  [138,]    0     0   1           0      0
#>  [139,]    0     1   0           1      1
#>  [140,]    0     0   1           1      1
#>  [141,]    0     0   1           0      1
#>  [142,]    1     0   0           1      1
#>  [143,]    0     0   1           1      1
#>  [144,]    0     0   1           0      1
#>  [145,]    0     0   1           0      1
#>  [146,]    0     0   1           1      0
#>  [147,]    1     0   0           1      0
#>  [148,]    1     0   0           0      1
#>  [149,]    0     1   0           0      0
#>  [150,]    0     0   1           0      0
#>  [151,]    0     0   1           0      1
#>  [152,]    0     0   1           0      1
#>  [153,]    0     0   1           1      1
#>  [154,]    0     1   0           1      1
#>  [155,]    0     0   1           0      0
#>  [156,]    0     1   0           0      1
#>  [157,]    0     0   1           1      1
#>  [158,]    0     1   0           1      0
#>  [159,]    0     0   1           1      1
#>  [160,]    0     1   0           1      0
#>  [161,]    0     0   1           1      0
#>  [162,]    0     0   1           1      0
#>  [163,]    0     1   0           1      0
#>  [164,]    1     0   0           0      1
#>  [165,]    0     0   1           1      0
#>  [166,]    0     1   0           1      0
#>  [167,]    0     0   1           0      1
#>  [168,]    0     1   0           1      1
#>  [169,]    0     0   1           1      1
#>  [170,]    1     0   0           0      1
#>  [171,]    1     0   0           0      1
#>  [172,]    0     1   0           1      0
#>  [173,]    0     1   0           0      1
#>  [174,]    0     0   1           0      1
#>  [175,]    0     1   0           1      0
#>  [176,]    1     0   0           1      1
#>  [177,]    0     1   0           0      0
#>  [178,]    1     0   0           1      0
#>  [179,]    0     0   1           1      0
#>  [180,]    1     0   0           0      1
#>  [181,]    0     0   1           0      1
#>  [182,]    0     1   0           0      0
#>  [183,]    0     1   0           1      1
#>  [184,]    0     0   1           1      0
#>  [185,]    0     1   0           0      0
#>  [186,]    1     0   0           0      1
#>  [187,]    1     0   0           0      0
#>  [188,]    0     0   1           1      0
#>  [189,]    0     0   1           0      1
#>  [190,]    1     0   0           1      0
#>  [191,]    0     1   0           0      0
#>  [192,]    1     0   0           1      0
#>  [193,]    0     1   0           1      0
#>  [194,]    0     1   0           0      0
#>  [195,]    0     1   0           1      0
#>  [196,]    0     0   1           0      1
#>  [197,]    0     1   0           1      1
#>  [198,]    0     0   1           0      0
#>  [199,]    1     0   0           0      0
#>  [200,]    0     1   0           1      0
#>  [201,]    0     1   0           0      1
#>  [202,]    0     1   0           1      1
#>  [203,]    1     0   0           0      0
#>  [204,]    0     0   1           1      1
#>  [205,]    0     0   1           1      0
#>  [206,]    0     0   1           0      0
#>  [207,]    0     0   1           0      1
#>  [208,]    0     1   0           1      1
#>  [209,]    0     0   1           0      0
#>  [210,]    0     0   1           1      1
#>  [211,]    0     1   0           0      1
#>  [212,]    0     0   1           1      0
#>  [213,]    1     0   0           0      1
#>  [214,]    0     0   1           1      1
#>  [215,]    0     1   0           0      0
#>  [216,]    1     0   0           1      0
#>  [217,]    1     0   0           0      1
#>  [218,]    0     0   1           1      1
#>  [219,]    1     0   0           0      0
#>  [220,]    1     0   0           1      0
#>  [221,]    0     1   0           0      1
#>  [222,]    0     0   1           0      0
#>  [223,]    0     0   1           1      1
#>  [224,]    0     1   0           1      1
#>  [225,]    0     1   0           1      1
#>  [226,]    0     1   0           0      0
#>  [227,]    1     0   0           0      1
#>  [228,]    0     1   0           1      0
#>  [229,]    0     1   0           1      0
#>  [230,]    0     1   0           1      1
#>  [231,]    0     0   1           0      0
#>  [232,]    0     0   1           1      1
#>  [233,]    0     0   1           1      1
#>  [234,]    0     0   1           1      0
#>  [235,]    1     0   0           0      1
#>  [236,]    0     0   1           0      1
#>  [237,]    1     0   0           1      0
#>  [238,]    0     1   0           1      1
#>  [239,]    0     1   0           1      0
#>  [240,]    0     0   1           0      1
#>  [241,]    0     1   0           1      0
#>  [242,]    0     0   1           0      0
#>  [243,]    1     0   0           1      0
#>  [244,]    0     0   1           0      1
#>  [245,]    1     0   0           1      0
#>  [246,]    1     0   0           1      1
#>  [247,]    1     0   0           0      1
#>  [248,]    0     0   1           0      0
#>  [249,]    0     0   1           0      0
#>  [250,]    0     1   0           0      1
#>  [251,]    0     0   1           1      0
#>  [252,]    0     1   0           1      1
#>  [253,]    1     0   0           0      1
#>  [254,]    0     0   1           1      0
#>  [255,]    0     1   0           1      1
#>  [256,]    0     1   0           0      1
#>  [257,]    1     0   0           0      0
#>  [258,]    1     0   0           1      0
#>  [259,]    1     0   0           0      0
#>  [260,]    0     1   0           1      1
#>  [261,]    1     0   0           1      1
#>  [262,]    0     1   0           0      0
#>  [263,]    0     1   0           1      1
#>  [264,]    0     0   1           0      1
#>  [265,]    0     0   1           0      0
#>  [266,]    0     0   1           1      1
#>  [267,]    0     1   0           0      1
#>  [268,]    1     0   0           1      1
#>  [269,]    0     0   1           0      1
#>  [270,]    0     0   1           1      1
#>  [271,]    0     0   1           0      1
#>  [272,]    1     0   0           0      0
#>  [273,]    1     0   0           0      0
#>  [274,]    1     0   0           0      0
#>  [275,]    0     1   0           0      1
#>  [276,]    0     0   1           1      1
#>  [277,]    0     1   0           1      1
#>  [278,]    1     0   0           1      1
#>  [279,]    0     0   1           0      0
#>  [280,]    0     0   1           0      1
#>  [281,]    0     0   1           0      0
#>  [282,]    0     0   1           0      0
#>  [283,]    0     0   1           1      1
#>  [284,]    0     0   1           0      0
#>  [285,]    0     0   1           0      1
#>  [286,]    0     0   1           1      0
#>  [287,]    0     1   0           0      0
#>  [288,]    0     0   1           0      1
#>  [289,]    0     0   1           0      0
#>  [290,]    0     0   1           1      1
#>  [291,]    1     0   0           0      1
#>  [292,]    1     0   0           1      0
#>  [293,]    1     0   0           1      1
#>  [294,]    1     0   0           0      1
#>  [295,]    0     0   1           0      0
#>  [296,]    0     0   1           0      0
#>  [297,]    1     0   0           0      0
#>  [298,]    1     0   0           1      1
#>  [299,]    0     0   1           1      0
#>  [300,]    1     0   0           0      1
#>  [301,]    1     0   0           0      0
#>  [302,]    0     1   0           0      0
#>  [303,]    0     1   0           0      1
#>  [304,]    0     0   1           1      0
#>  [305,]    0     0   1           1      0
#>  [306,]    0     1   0           0      0
#>  [307,]    0     1   0           1      1
#>  [308,]    0     0   1           1      0
#>  [309,]    0     1   0           1      1
#>  [310,]    1     0   0           0      1
#>  [311,]    0     0   1           1      0
#>  [312,]    0     0   1           0      0
#>  [313,]    1     0   0           1      0
#>  [314,]    1     0   0           0      1
#>  [315,]    0     0   1           1      0
#>  [316,]    0     0   1           0      1
#>  [317,]    1     0   0           1      0
#>  [318,]    1     0   0           0      0
#>  [319,]    0     0   1           1      0
#>  [320,]    0     0   1           0      1
#>  [321,]    1     0   0           1      0
#>  [322,]    0     0   1           0      1
#>  [323,]    0     0   1           1      1
#>  [324,]    1     0   0           1      1
#>  [325,]    0     0   1           1      1
#>  [326,]    0     0   1           1      1
#>  [327,]    1     0   0           0      0
#>  [328,]    0     1   0           0      1
#>  [329,]    0     0   1           1      1
#>  [330,]    1     0   0           0      0
#>  [331,]    1     0   0           0      0
#>  [332,]    0     0   1           1      0
#>  [333,]    0     1   0           1      1
#>  [334,]    0     1   0           0      1
#>  [335,]    0     1   0           0      0
#>  [336,]    0     1   0           0      0
#>  [337,]    1     0   0           0      0
#>  [338,]    0     1   0           1      0
#>  [339,]    0     0   1           1      1
#>  [340,]    0     1   0           1      0
#>  [341,]    0     0   1           1      0
#>  [342,]    0     1   0           0      1
#>  [343,]    0     0   1           0      1
#>  [344,]    1     0   0           1      0
#>  [345,]    0     0   1           0      1
#>  [346,]    0     0   1           0      1
#>  [347,]    0     1   0           0      1
#>  [348,]    0     0   1           1      0
#>  [349,]    1     0   0           1      0
#>  [350,]    0     1   0           1      1
#>  [351,]    0     0   1           1      0
#>  [352,]    0     1   0           1      1
#>  [353,]    1     0   0           1      1
#>  [354,]    0     1   0           1      1
#>  [355,]    0     1   0           0      1
#>  [356,]    0     0   1           1      1
#>  [357,]    1     0   0           1      1
#>  [358,]    0     0   1           0      0
#>  [359,]    0     0   1           1      0
#>  [360,]    1     0   0           0      1
#>  [361,]    0     0   1           1      0
#>  [362,]    0     0   1           1      1
#>  [363,]    1     0   0           1      1
#>  [364,]    0     0   1           1      1
#>  [365,]    0     0   1           0      0
#>  [366,]    0     1   0           1      0
#>  [367,]    0     1   0           1      0
#>  [368,]    0     1   0           1      0
#>  [369,]    0     0   1           1      1
#>  [370,]    1     0   0           1      1
#>  [371,]    0     1   0           0      1
#>  [372,]    0     1   0           0      1
#>  [373,]    1     0   0           0      1
#>  [374,]    1     0   0           1      0
#>  [375,]    0     0   1           0      1
#>  [376,]    0     0   1           1      1
#>  [377,]    1     0   0           0      0
#>  [378,]    0     1   0           1      1
#>  [379,]    1     0   0           1      1
#>  [380,]    0     0   1           0      0
#>  [381,]    1     0   0           1      1
#>  [382,]    0     1   0           0      0
#>  [383,]    1     0   0           1      1
#>  [384,]    0     0   1           1      0
#>  [385,]    0     0   1           0      0
#>  [386,]    0     0   1           1      1
#>  [387,]    0     0   1           1      0
#>  [388,]    1     0   0           0      0
#>  [389,]    0     0   1           0      1
#>  [390,]    1     0   0           1      1
#>  [391,]    1     0   0           1      1
#>  [392,]    0     0   1           1      0
#>  [393,]    0     0   1           0      0
#>  [394,]    0     0   1           0      1
#>  [395,]    0     0   1           1      0
#>  [396,]    0     0   1           1      0
#>  [397,]    1     0   0           1      0
#>  [398,]    1     0   0           1      0
#>  [399,]    0     0   1           1      0
#>  [400,]    1     0   0           1      1
#>  [401,]    0     1   0           1      0
#>  [402,]    0     0   1           1      0
#>  [403,]    0     1   0           0      1
#>  [404,]    1     0   0           0      1
#>  [405,]    0     0   1           0      0
#>  [406,]    0     0   1           1      0
#>  [407,]    0     1   0           0      1
#>  [408,]    0     0   1           0      0
#>  [409,]    1     0   0           0      0
#>  [410,]    0     0   1           0      1
#>  [411,]    0     1   0           1      1
#>  [412,]    0     0   1           1      0
#>  [413,]    0     0   1           0      1
#>  [414,]    1     0   0           0      1
#>  [415,]    0     0   1           1      1
#>  [416,]    0     1   0           0      0
#>  [417,]    0     0   1           1      1
#>  [418,]    0     1   0           0      0
#>  [419,]    0     1   0           0      0
#>  [420,]    0     1   0           1      1
#>  [421,]    0     0   1           0      1
#>  [422,]    1     0   0           1      0
#>  [423,]    0     1   0           1      1
#>  [424,]    0     0   1           1      1
#>  [425,]    0     0   1           0      0
#>  [426,]    0     0   1           0      1
#>  [427,]    1     0   0           1      0
#>  [428,]    1     0   0           0      0
#>  [429,]    0     1   0           1      1
#>  [430,]    0     0   1           0      1
#>  [431,]    1     0   0           0      0
#>  [432,]    0     0   1           1      1
#>  [433,]    0     1   0           1      0
#>  [434,]    0     0   1           0      1
#>  [435,]    0     1   0           0      0
#>  [436,]    0     0   1           0      1
#>  [437,]    0     0   1           0      0
#>  [438,]    0     0   1           1      0
#>  [439,]    0     1   0           0      1
#>  [440,]    1     0   0           1      1
#>  [441,]    0     0   1           1      1
#>  [442,]    1     0   0           0      1
#>  [443,]    1     0   0           0      0
#>  [444,]    1     0   0           1      1
#>  [445,]    0     0   1           1      1
#>  [446,]    0     0   1           0      1
#>  [447,]    0     1   0           1      0
#>  [448,]    0     1   0           0      0
#>  [449,]    0     0   1           0      1
#>  [450,]    0     0   1           0      1
#>  [451,]    0     0   1           1      0
#>  [452,]    0     0   1           0      0
#>  [453,]    1     0   0           0      0
#>  [454,]    1     0   0           1      0
#>  [455,]    1     0   0           0      1
#>  [456,]    0     0   1           0      0
#>  [457,]    0     0   1           1      1
#>  [458,]    1     0   0           0      1
#>  [459,]    0     1   0           0      0
#>  [460,]    0     1   0           0      0
#>  [461,]    0     0   1           1      1
#>  [462,]    0     1   0           1      1
#>  [463,]    0     0   1           0      1
#>  [464,]    0     0   1           0      1
#>  [465,]    0     1   0           1      1
#>  [466,]    0     0   1           1      1
#>  [467,]    1     0   0           1      1
#>  [468,]    0     0   1           0      1
#>  [469,]    1     0   0           1      1
#>  [470,]    0     1   0           0      1
#>  [471,]    1     0   0           1      0
#>  [472,]    0     1   0           1      1
#>  [473,]    0     1   0           0      1
#>  [474,]    0     1   0           1      0
#>  [475,]    0     0   1           0      1
#>  [476,]    0     0   1           0      0
#>  [477,]    1     0   0           0      1
#>  [478,]    0     0   1           0      1
#>  [479,]    1     0   0           1      1
#>  [480,]    0     0   1           0      0
#>  [481,]    0     1   0           0      1
#>  [482,]    1     0   0           0      0
#>  [483,]    0     0   1           1      1
#>  [484,]    0     0   1           0      1
#>  [485,]    0     1   0           1      0
#>  [486,]    0     1   0           0      1
#>  [487,]    1     0   0           0      0
#>  [488,]    1     0   0           1      0
#>  [489,]    0     0   1           1      1
#>  [490,]    0     1   0           0      0
#>  [491,]    1     0   0           0      0
#>  [492,]    0     0   1           1      0
#>  [493,]    1     0   0           0      0
#>  [494,]    0     0   1           0      0
#>  [495,]    0     0   1           1      1
#>  [496,]    0     1   0           0      0
#>  [497,]    1     0   0           1      1
#>  [498,]    0     0   1           0      0
#>  [499,]    0     0   1           0      0
#>  [500,]    0     0   1           1      0
#>  [501,]    0     0   1           1      1
#>  [502,]    0     0   1           0      0
#>  [503,]    0     0   1           0      1
#>  [504,]    0     1   0           0      0
#>  [505,]    0     1   0           1      0
#>  [506,]    1     0   0           1      0
#>  [507,]    0     0   1           0      0
#>  [508,]    0     1   0           1      1
#>  [509,]    0     0   1           0      1
#>  [510,]    1     0   0           0      1
#>  [511,]    0     0   1           1      1
#>  [512,]    1     0   0           1      0
#>  [513,]    0     0   1           1      0
#>  [514,]    0     1   0           1      1
#>  [515,]    1     0   0           1      0
#>  [516,]    0     0   1           1      0
#>  [517,]    1     0   0           1      1
#>  [518,]    0     0   1           0      0
#>  [519,]    0     0   1           1      0
#>  [520,]    0     0   1           0      0
#>  [521,]    0     0   1           1      1
#>  [522,]    0     1   0           0      1
#>  [523,]    0     1   0           1      0
#>  [524,]    1     0   0           1      1
#>  [525,]    0     0   1           1      1
#>  [526,]    1     0   0           0      0
#>  [527,]    0     0   1           1      0
#>  [528,]    1     0   0           0      0
#>  [529,]    0     1   0           0      0
#>  [530,]    0     1   0           0      0
#>  [531,]    1     0   0           1      0
#>  [532,]    0     0   1           1      0
#>  [533,]    1     0   0           1      1
#>  [534,]    0     0   1           1      1
#>  [535,]    1     0   0           1      0
#>  [536,]    0     1   0           0      0
#>  [537,]    0     0   1           1      0
#>  [538,]    1     0   0           1      0
#>  [539,]    0     0   1           0      1
#>  [540,]    1     0   0           1      0
#>  [541,]    0     0   1           0      0
#>  [542,]    1     0   0           0      0
#>  [543,]    1     0   0           1      0
#>  [544,]    0     0   1           1      1
#>  [545,]    1     0   0           1      0
#>  [546,]    1     0   0           0      1
#>  [547,]    0     0   1           0      1
#>  [548,]    0     0   1           0      0
#>  [549,]    0     0   1           1      1
#>  [550,]    0     1   0           1      0
#>  [551,]    1     0   0           1      1
#>  [552,]    0     1   0           1      0
#>  [553,]    0     1   0           1      0
#>  [554,]    0     1   0           0      1
#>  [555,]    0     1   0           1      1
#>  [556,]    0     1   0           1      1
#>  [557,]    0     0   1           0      1
#>  [558,]    0     0   1           0      1
#>  [559,]    1     0   0           1      0
#>  [560,]    1     0   0           0      1
#>  [561,]    0     0   1           1      1
#>  [562,]    0     1   0           0      1
#>  [563,]    0     1   0           1      0
#>  [564,]    0     0   1           1      1
#>  [565,]    1     0   0           0      1
#>  [566,]    0     1   0           1      1
#>  [567,]    1     0   0           0      0
#>  [568,]    0     1   0           1      1
#>  [569,]    1     0   0           1      0
#>  [570,]    0     1   0           0      1
#>  [571,]    0     1   0           1      1
#>  [572,]    1     0   0           1      1
#>  [573,]    1     0   0           0      1
#>  [574,]    0     1   0           1      1
#>  [575,]    0     0   1           1      1
#>  [576,]    0     1   0           1      1
#>  [577,]    0     1   0           1      0
#>  [578,]    1     0   0           0      0
#>  [579,]    1     0   0           1      1
#>  [580,]    1     0   0           1      0
#>  [581,]    0     0   1           1      1
#>  [582,]    1     0   0           0      1
#>  [583,]    0     0   1           0      0
#>  [584,]    0     0   1           0      0
#>  [585,]    0     1   0           1      0
#>  [586,]    1     0   0           1      0
#>  [587,]    0     1   0           1      1
#>  [588,]    0     1   0           0      0
#>  [589,]    0     0   1           0      0
#>  [590,]    1     0   0           0      0
#>  [591,]    0     0   1           0      1
#>  [592,]    0     0   1           1      1
#>  [593,]    1     0   0           1      0
#>  [594,]    0     1   0           0      1
#>  [595,]    0     0   1           1      0
#>  [596,]    1     0   0           0      0
#>  [597,]    0     0   1           1      1
#>  [598,]    0     0   1           0      1
#>  [599,]    0     1   0           1      0
#>  [600,]    0     0   1           1      1
#>  [601,]    0     0   1           0      0
#>  [602,]    0     0   1           0      1
#>  [603,]    1     0   0           1      1
#>  [604,]    0     0   1           0      0
#>  [605,]    1     0   0           1      1
#>  [606,]    0     0   1           0      0
#>  [607,]    0     0   1           1      0
#>  [608,]    0     1   0           0      0
#>  [609,]    0     0   1           0      1
#>  [610,]    0     0   1           1      1
#>  [611,]    1     0   0           0      1
#>  [612,]    0     1   0           0      0
#>  [613,]    1     0   0           1      0
#>  [614,]    1     0   0           0      1
#>  [615,]    1     0   0           1      0
#>  [616,]    1     0   0           1      1
#>  [617,]    0     0   1           1      0
#>  [618,]    1     0   0           1      0
#>  [619,]    1     0   0           1      1
#>  [620,]    1     0   0           1      1
#>  [621,]    0     0   1           0      0
#>  [622,]    0     0   1           1      0
#>  [623,]    0     0   1           0      0
#>  [624,]    0     0   1           0      1
#>  [625,]    1     0   0           1      1
#>  [626,]    0     0   1           0      1
#>  [627,]    0     0   1           0      1
#>  [628,]    1     0   0           0      0
#>  [629,]    1     0   0           0      0
#>  [630,]    0     1   0           1      1
#>  [631,]    1     0   0           0      1
#>  [632,]    0     0   1           0      1
#>  [633,]    0     0   1           0      1
#>  [634,]    1     0   0           1      0
#>  [635,]    1     0   0           1      0
#>  [636,]    1     0   0           1      0
#>  [637,]    0     1   0           0      1
#>  [638,]    0     1   0           1      1
#>  [639,]    1     0   0           0      0
#>  [640,]    1     0   0           0      1
#>  [641,]    0     0   1           0      1
#>  [642,]    0     0   1           1      0
#>  [643,]    0     1   0           0      0
#>  [644,]    1     0   0           0      0
#>  [645,]    0     0   1           0      1
#>  [646,]    0     0   1           1      0
#>  [647,]    0     1   0           1      1
#>  [648,]    1     0   0           1      0
#>  [649,]    1     0   0           0      0
#>  [650,]    0     1   0           1      0
#>  [651,]    0     1   0           1      1
#>  [652,]    1     0   0           1      1
#>  [653,]    1     0   0           0      0
#>  [654,]    1     0   0           1      1
#>  [655,]    0     0   1           0      0
#>  [656,]    0     1   0           1      1
#>  [657,]    0     0   1           0      0
#>  [658,]    1     0   0           0      1
#>  [659,]    1     0   0           1      1
#>  [660,]    0     0   1           1      0
#>  [661,]    1     0   0           0      1
#>  [662,]    0     0   1           0      1
#>  [663,]    0     0   1           1      0
#>  [664,]    1     0   0           0      1
#>  [665,]    0     0   1           1      0
#>  [666,]    0     1   0           1      1
#>  [667,]    0     0   1           1      1
#>  [668,]    0     0   1           0      1
#>  [669,]    0     1   0           0      1
#>  [670,]    0     0   1           1      0
#>  [671,]    0     0   1           0      1
#>  [672,]    0     1   0           1      1
#>  [673,]    0     1   0           0      1
#>  [674,]    1     0   0           0      1
#>  [675,]    0     0   1           0      0
#>  [676,]    0     1   0           0      1
#>  [677,]    0     1   0           1      1
#>  [678,]    1     0   0           1      1
#>  [679,]    1     0   0           0      1
#>  [680,]    1     0   0           0      0
#>  [681,]    1     0   0           1      0
#>  [682,]    1     0   0           1      1
#>  [683,]    0     1   0           1      0
#>  [684,]    0     1   0           0      1
#>  [685,]    0     0   1           0      0
#>  [686,]    0     1   0           0      1
#>  [687,]    0     0   1           0      0
#>  [688,]    0     0   1           1      1
#>  [689,]    0     1   0           1      1
#>  [690,]    0     1   0           1      1
#>  [691,]    0     0   1           0      0
#>  [692,]    0     0   1           0      0
#>  [693,]    0     1   0           0      0
#>  [694,]    0     0   1           0      1
#>  [695,]    0     0   1           0      0
#>  [696,]    0     1   0           1      0
#>  [697,]    0     0   1           1      0
#>  [698,]    1     0   0           0      1
#>  [699,]    0     0   1           1      0
#>  [700,]    0     0   1           1      1
#>  [701,]    1     0   0           0      0
#>  [702,]    0     0   1           0      1
#>  [703,]    0     1   0           0      0
#>  [704,]    0     1   0           1      0
#>  [705,]    0     0   1           0      0
#>  [706,]    0     0   1           0      0
#>  [707,]    0     1   0           0      1
#>  [708,]    0     1   0           1      0
#>  [709,]    0     1   0           1      1
#>  [710,]    0     1   0           1      1
#>  [711,]    1     0   0           0      1
#>  [712,]    0     1   0           1      1
#>  [713,]    0     0   1           0      0
#>  [714,]    0     0   1           0      1
#>  [715,]    0     0   1           0      1
#>  [716,]    1     0   0           0      0
#>  [717,]    0     0   1           1      0
#>  [718,]    0     1   0           0      0
#>  [719,]    1     0   0           1      1
#>  [720,]    1     0   0           0      0
#>  [721,]    1     0   0           1      0
#>  [722,]    0     1   0           1      0
#>  [723,]    1     0   0           0      0
#>  [724,]    1     0   0           1      0
#>  [725,]    0     1   0           0      1
#>  [726,]    1     0   0           1      1
#>  [727,]    0     0   1           1      1
#>  [728,]    0     1   0           0      1
#>  [729,]    1     0   0           1      0
#>  [730,]    0     1   0           1      0
#>  [731,]    1     0   0           0      1
#>  [732,]    0     0   1           0      0
#>  [733,]    0     0   1           0      1
#>  [734,]    1     0   0           1      0
#>  [735,]    1     0   0           0      0
#>  [736,]    0     0   1           0      0
#>  [737,]    0     0   1           0      0
#>  [738,]    1     0   0           1      1
#>  [739,]    0     0   1           0      1
#>  [740,]    0     0   1           0      1
#>  [741,]    0     1   0           1      1
#>  [742,]    0     0   1           1      0
#>  [743,]    0     1   0           1      0
#>  [744,]    1     0   0           0      1
#>  [745,]    0     0   1           0      1
#>  [746,]    0     1   0           0      1
#>  [747,]    1     0   0           1      0
#>  [748,]    0     0   1           0      0
#>  [749,]    0     0   1           0      1
#>  [750,]    0     0   1           0      1
#>  [751,]    0     0   1           0      0
#>  [752,]    0     1   0           0      1
#>  [753,]    0     1   0           1      1
#>  [754,]    0     1   0           0      1
#>  [755,]    1     0   0           1      0
#>  [756,]    0     0   1           1      0
#>  [757,]    0     0   1           0      1
#>  [758,]    0     1   0           0      0
#>  [759,]    0     0   1           1      1
#>  [760,]    0     1   0           1      0
#>  [761,]    0     1   0           1      1
#>  [762,]    0     1   0           1      0
#>  [763,]    0     0   1           0      1
#>  [764,]    0     0   1           0      0
#>  [765,]    0     1   0           1      1
#>  [766,]    0     1   0           0      0
#>  [767,]    0     1   0           0      0
#>  [768,]    1     0   0           0      1
#>  [769,]    1     0   0           0      1
#>  [770,]    0     1   0           1      1
#>  [771,]    1     0   0           0      1
#>  [772,]    0     0   1           1      0
#>  [773,]    0     0   1           1      0
#>  [774,]    0     0   1           1      0
#>  [775,]    0     1   0           1      1
#>  [776,]    0     1   0           0      1
#>  [777,]    1     0   0           1      1
#>  [778,]    0     0   1           1      1
#>  [779,]    1     0   0           1      0
#>  [780,]    0     0   1           0      1
#>  [781,]    0     1   0           0      0
#>  [782,]    1     0   0           1      0
#>  [783,]    0     1   0           1      0
#>  [784,]    0     0   1           0      0
#>  [785,]    0     0   1           1      1
#>  [786,]    0     0   1           0      0
#>  [787,]    1     0   0           0      0
#>  [788,]    1     0   0           1      1
#>  [789,]    0     1   0           1      1
#>  [790,]    0     0   1           0      0
#>  [791,]    0     1   0           0      1
#>  [792,]    1     0   0           1      0
#>  [793,]    0     1   0           1      1
#>  [794,]    0     0   1           0      0
#>  [795,]    0     0   1           0      0
#>  [796,]    0     0   1           0      1
#>  [797,]    0     0   1           0      0
#>  [798,]    0     1   0           1      1
#>  [799,]    0     0   1           0      0
#>  [800,]    0     0   1           1      1
#>  [801,]    0     0   1           0      1
#>  [802,]    0     0   1           0      0
#>  [803,]    1     0   0           1      0
#>  [804,]    0     1   0           0      1
#>  [805,]    0     1   0           0      0
#>  [806,]    1     0   0           0      0
#>  [807,]    0     1   0           0      0
#>  [808,]    0     0   1           1      0
#>  [809,]    0     1   0           0      1
#>  [810,]    0     0   1           0      0
#>  [811,]    1     0   0           0      0
#>  [812,]    1     0   0           1      0
#>  [813,]    0     0   1           1      1
#>  [814,]    0     0   1           0      1
#>  [815,]    0     0   1           1      0
#>  [816,]    0     1   0           0      1
#>  [817,]    0     1   0           1      0
#>  [818,]    0     1   0           0      0
#>  [819,]    0     0   1           1      1
#>  [820,]    0     1   0           1      0
#>  [821,]    1     0   0           0      0
#>  [822,]    1     0   0           0      1
#>  [823,]    0     0   1           0      1
#>  [824,]    0     1   0           1      0
#>  [825,]    0     0   1           0      1
#>  [826,]    0     0   1           0      0
#>  [827,]    1     0   0           0      1
#>  [828,]    0     0   1           0      0
#>  [829,]    1     0   0           0      0
#>  [830,]    0     1   0           0      0
#>  [831,]    1     0   0           1      0
#>  [832,]    0     1   0           1      1
#>  [833,]    0     1   0           1      1
#>  [834,]    1     0   0           0      0
#>  [835,]    0     1   0           1      0
#>  [836,]    0     0   1           0      0
#>  [837,]    0     0   1           0      0
#>  [838,]    1     0   0           1      1
#>  [839,]    1     0   0           1      1
#>  [840,]    0     1   0           1      0
#>  [841,]    0     0   1           0      0
#>  [842,]    0     1   0           1      0
#>  [843,]    0     0   1           1      0
#>  [844,]    1     0   0           0      0
#>  [845,]    0     0   1           1      1
#>  [846,]    0     0   1           0      1
#>  [847,]    0     0   1           1      1
#>  [848,]    0     0   1           0      0
#>  [849,]    1     0   0           0      1
#>  [850,]    1     0   0           1      1
#>  [851,]    1     0   0           1      1
#>  [852,]    0     1   0           0      0
#>  [853,]    0     1   0           0      1
#>  [854,]    1     0   0           0      1
#>  [855,]    0     0   1           1      0
#>  [856,]    0     0   1           0      1
#>  [857,]    1     0   0           0      0
#>  [858,]    0     1   0           0      0
#>  [859,]    0     0   1           0      1
#>  [860,]    0     1   0           1      1
#>  [861,]    0     1   0           0      1
#>  [862,]    0     1   0           0      0
#>  [863,]    0     0   1           1      0
#>  [864,]    0     0   1           1      0
#>  [865,]    0     0   1           0      0
#>  [866,]    0     1   0           1      1
#>  [867,]    0     1   0           0      1
#>  [868,]    0     0   1           1      0
#>  [869,]    0     0   1           0      1
#>  [870,]    1     0   0           1      1
#>  [871,]    0     1   0           1      1
#>  [872,]    1     0   0           1      0
#>  [873,]    0     1   0           0      0
#>  [874,]    0     0   1           1      1
#>  [875,]    0     1   0           1      1
#>  [876,]    1     0   0           0      1
#>  [877,]    0     1   0           1      0
#>  [878,]    0     0   1           0      1
#>  [879,]    1     0   0           1      0
#>  [880,]    0     1   0           1      0
#>  [881,]    0     0   1           1      0
#>  [882,]    1     0   0           1      1
#>  [883,]    1     0   0           0      0
#>  [884,]    0     1   0           1      0
#>  [885,]    0     0   1           1      1
#>  [886,]    0     0   1           1      0
#>  [887,]    0     0   1           0      0
#>  [888,]    0     1   0           0      1
#>  [889,]    1     0   0           0      0
#>  [890,]    0     1   0           1      1
#>  [891,]    0     1   0           0      1
#>  [892,]    1     0   0           0      1
#>  [893,]    1     0   0           0      1
#>  [894,]    0     0   1           1      0
#>  [895,]    0     0   1           0      0
#>  [896,]    0     0   1           1      1
#>  [897,]    1     0   0           0      1
#>  [898,]    1     0   0           1      0
#>  [899,]    0     1   0           1      0
#>  [900,]    1     0   0           1      0
#>  [901,]    0     1   0           0      1
#>  [902,]    0     1   0           0      0
#>  [903,]    0     1   0           0      0
#>  [904,]    1     0   0           0      0
#>  [905,]    1     0   0           0      1
#>  [906,]    1     0   0           0      1
#>  [907,]    0     0   1           0      0
#>  [908,]    0     0   1           1      0
#>  [909,]    0     0   1           0      1
#>  [910,]    1     0   0           0      0
#>  [911,]    0     1   0           0      1
#>  [912,]    0     0   1           0      0
#>  [913,]    1     0   0           1      0
#>  [914,]    0     1   0           1      0
#>  [915,]    0     1   0           0      0
#>  [916,]    0     0   1           0      0
#>  [917,]    0     0   1           1      0
#>  [918,]    1     0   0           0      1
#>  [919,]    0     0   1           1      0
#>  [920,]    0     0   1           0      0
#>  [921,]    1     0   0           1      1
#>  [922,]    0     1   0           0      0
#>  [923,]    0     1   0           1      1
#>  [924,]    0     0   1           0      0
#>  [925,]    0     1   0           1      0
#>  [926,]    0     0   1           0      0
#>  [927,]    1     0   0           0      0
#>  [928,]    1     0   0           0      1
#>  [929,]    0     0   1           1      1
#>  [930,]    1     0   0           1      1
#>  [931,]    0     1   0           1      1
#>  [932,]    1     0   0           1      0
#>  [933,]    0     0   1           1      1
#>  [934,]    0     1   0           1      0
#>  [935,]    1     0   0           0      1
#>  [936,]    1     0   0           1      1
#>  [937,]    0     0   1           0      0
#>  [938,]    1     0   0           0      1
#>  [939,]    0     1   0           1      0
#>  [940,]    0     0   1           1      1
#>  [941,]    0     1   0           0      1
#>  [942,]    0     0   1           1      1
#>  [943,]    0     0   1           0      0
#>  [944,]    0     1   0           1      1
#>  [945,]    0     0   1           0      0
#>  [946,]    0     1   0           0      1
#>  [947,]    0     0   1           1      1
#>  [948,]    0     1   0           1      1
#>  [949,]    0     0   1           0      0
#>  [950,]    1     0   0           1      0
#>  [951,]    0     0   1           1      1
#>  [952,]    0     0   1           0      1
#>  [953,]    1     0   0           1      0
#>  [954,]    0     1   0           1      1
#>  [955,]    0     0   1           0      0
#>  [956,]    0     1   0           0      1
#>  [957,]    0     1   0           1      1
#>  [958,]    0     1   0           0      0
#>  [959,]    0     1   0           1      1
#>  [960,]    0     1   0           1      0
#>  [961,]    0     0   1           1      1
#>  [962,]    1     0   0           1      0
#>  [963,]    0     1   0           1      1
#>  [964,]    0     1   0           1      0
#>  [965,]    0     0   1           1      1
#>  [966,]    1     0   0           0      1
#>  [967,]    0     1   0           0      1
#>  [968,]    1     0   0           1      0
#>  [969,]    0     0   1           0      1
#>  [970,]    1     0   0           1      1
#>  [971,]    0     0   1           0      1
#>  [972,]    0     0   1           0      0
#>  [973,]    0     1   0           0      1
#>  [974,]    1     0   0           1      0
#>  [975,]    0     0   1           1      1
#>  [976,]    0     0   1           0      0
#>  [977,]    0     1   0           1      1
#>  [978,]    0     1   0           0      0
#>  [979,]    1     0   0           1      1
#>  [980,]    1     0   0           1      1
#>  [981,]    0     1   0           1      1
#>  [982,]    1     0   0           0      1
#>  [983,]    1     0   0           1      0
#>  [984,]    0     0   1           1      0
#>  [985,]    0     0   1           1      1
#>  [986,]    0     0   1           0      1
#>  [987,]    0     1   0           1      0
#>  [988,]    0     0   1           1      0
#>  [989,]    0     0   1           1      1
#>  [990,]    0     1   0           0      0
#>  [991,]    0     1   0           1      1
#>  [992,]    0     1   0           1      0
#>  [993,]    1     0   0           0      1
#>  [994,]    0     1   0           1      0
#>  [995,]    0     0   1           0      1
#>  [996,]    1     0   0           1      1
#>  [997,]    0     0   1           1      0
#>  [998,]    0     0   1           1      1
#>  [999,]    0     1   0           1      0
#> [1000,]    0     0   1           1      1